moes formatting functions into formatter file
