+++ lib/books.js
@@ -1,7 +1,7 @@
 module.exports = {makeCall}
 requests = require('./requests')
+formatter = require('./formatter')
 
-const request = require('request');
 const tableHeaders = '<thead><tr><th scope="col">Price (BTC)</th>'
                     + '<th scope="col">Bittrex</th>'
                     + '<th scope="col">Poloniex</th>'
@@ -17,7 +17,7 @@ function makeCall(){
   Promise.all([requests.requestAsync(url1), requests.requestAsync(url2)])
       .then(function(allData) {
         var lowerCaseData = JSON.stringify(allData).toLowerCase()
-        var formattedData = formatData(JSON.parse(lowerCaseData))
+        var formattedData = formatter.formatData(JSON.parse(lowerCaseData))
         data = generateOrderBooks(formattedData)
       });
       return JSON.stringify(data)
@@ -86,32 +86,4 @@ function combineData(data, orderType){
     })
   })
   return rowData
-}
-
-
-
-function formatData(dataCollection){
-  var bittrexData = formatBittrex(dataCollection[0])
-  var poloniexData = formatPoloniex(dataCollection[1])
-  return [bittrexData, poloniexData]
-}
-
-function formatBittrex(data){
-  return {name: "bittrex",
-          bids: data["result"]["buy"].slice(0, 50),
-          asks: data["result"]["sell"].slice(0, 50)}
-}
-
-function formatPoloniex(data){
-    var bids = data["bids"].map(function(bid){
-      return {rate: bid[0],
-              quantity: bid[1]}
-    })
-    var asks = data["asks"].map(function(bid){
-      return {rate: bid[0],
-              quantity: bid[1]}
-    })
-    return {name: "poloniex",
-            bids: bids,
-            asks: asks}
 }
